+++ src/Queue/AbstractQueue.php
@@ -14,7 +14,12 @@ declare(strict_types=1);
 
 namespace Berlioz\QueueManager\Queue;
 
-abstract readonly class AbstractQueue implements QueueInterface
+use DateInterval;
+use DateTimeImmutable;
+use DateTimeInterface;
+use Psr\Clock\ClockInterface;
+
+abstract readonly class AbstractQueue implements QueueInterface, ClockInterface
 {
     public function __construct(
         protected string $name = 'default',
@@ -24,8 +29,68 @@ abstract readonly class AbstractQueue implements QueueInterface
     /**
      * @inheritDoc
      */
+    public function now(): DateTimeImmutable
+    {
+        return new DateTimeImmutable();
+    }
+
+    /**
+     * @inheritDoc
+     */
     public function getName(): string
     {
         return $this->name;
+    }
+
+    /**
+     * Get available date time.
+     *
+     * @param DateTimeInterface|DateInterval|int $delay
+     *
+     * @return DateTimeInterface
+     */
+    protected function getAvailableDateTime(DateTimeInterface|DateInterval|int $delay): DateTimeInterface
+    {
+        $now = $this->now();
+
+        if ($delay instanceof DateTimeInterface) {
+            if ($now->diff($delay)->invert === 1) {
+                return $now;
+            }
+
+            return $delay;
+        }
+
+        if ($delay instanceof DateInterval) {
+            if ($delay->invert === 1) {
+                return $now;
+            }
+
+            return $now->add($delay);
+        }
+
+        if ($delay < 0) {
+            return $now;
+        }
+
+        return $now->add(new DateInterval('PT' . $delay . 'S'));
+    }
+
+    /**
+     * Get delay in seconds.
+     *
+     * @param DateTimeInterface $dateTime
+     *
+     * @return int
+     */
+    protected function getDelayInSeconds(DateTimeInterface $dateTime): int
+    {
+        $diff = $this->now()->diff($dateTime);
+
+        return
+            $diff->days * (24 * 60 * 60) +
+            $diff->h * (60 * 60) +
+            $diff->i * (60) +
+            $diff->s;
     }
 }