+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta7] - 2025-01-27
+
+### Changed
+
+- Methods `QueueInterface::push()` and `QueueInterface::pushRaw()`, accept `\DateTimeInterface` or `\DateInterval` for `$delay` parameter
+
+### Fixed
+
+- Attempts value do not increased after consume
+
 ## [1.0.0-beta6] - 2025-01-23
 
 ### Fixed